+++ config/fake-env/webextension-page.js
@@ -36,6 +36,9 @@ window.browser = {
     removeAll () { return Promise.resolve() },
     create () { return Promise.resolve() },
   },
+  cookies: {
+    set: () => Promise.resolve()
+  },
   extension: {
     inIncognitoContext: false,
   },