+++ src/components/dictionaries/hjdict/engine.ts
@@ -7,7 +7,6 @@ import {
   isContainDeutsch,
   isContainSpanish,
 } from '@/_helpers/lang-check'
-import { fetchDirtyDOM } from '@/_helpers/fetch-dom'
 import {
   HTMLString,
   getInnerHTMLBuilder,
@@ -40,10 +39,34 @@ export type HjdictResult = HjdictResultLex | HjdictResultRelated
 
 type HjdictSearchResult = DictSearchResult<HjdictResult>
 
-export const search: SearchFunction<HjdictSearchResult> = (
+export const search: SearchFunction<HjdictSearchResult> = async (
   text, config, payload
 ) => {
-  return fetchDirtyDOM(getSrcPage(text, config))
+  const cookies = {
+    HJ_SITEID: 3,
+    HJ_UID: getUUID(),
+    HJ_SID: getUUID(),
+    HJ_SSID: getUUID(),
+    HJID: 0,
+    HJ_VT: 2,
+    HJ_SST: 1,
+    HJ_CSST: 1,
+    HJ_ST: 1,
+    HJ_CST: 1,
+    HJ_T: +new Date(),
+    _: getUUID(16),
+  }
+
+  await Promise.all(Object.keys(cookies).map(name =>
+    browser.cookies.set({
+      url: 'https://www.hjdict.com',
+      domain: 'hjdict.com',
+      name,
+      value: String(cookies[name]),
+    })
+  ))
+
+  return xhrDirtyDOM(getSrcPage(text, config))
     .catch(handleNetWorkError)
     .then(doc => handleDOM(doc, config.dicts.all.hjdict.options))
 }
@@ -99,6 +122,30 @@ function handleDOM (
     : handleNoResult()
 }
 
+/**
+ * Firefox adds 'Origin' field with `fetch` which would be rejected by the server.
+ */
+function xhrDirtyDOM (url: string): Promise<Document> {
+  return new Promise((resolve, reject) => {
+    const xhr = new XMLHttpRequest()
+    xhr.open('GET', url, true)
+    xhr.responseType = 'document'
+    xhr.withCredentials = true
+    xhr.onload = () => {
+      console.log(xhr.readyState)
+      if (xhr.readyState === xhr.DONE && xhr.status >= 200 && xhr.status < 300) {
+        if (xhr.responseXML) {
+          resolve(xhr.responseXML)
+        } else {
+          reject(xhr)
+        }
+      }
+    }
+    xhr.onerror = err => reject(err)
+    xhr.send(null)
+  })
+}
+
 function getLangCode (text: string, config: AppConfig): string {
   // ü
   if (/\u00fc/i.test(text)) {
@@ -144,4 +191,22 @@ function getLangCode (text: string, config: AppConfig): string {
   }
 
   return 'w'
+}
+
+function getUUID (e?: number): string {
+  let t = arguments.length > 1 && void 0 !== arguments[1] ? arguments[1] : 16
+  let n = ''
+  if ('number' === typeof e) {
+    for (let i = 0; i < e; i++) {
+      let r = Math.floor(10 * Math.random())
+      n += r % 2 === 0 ? 'x' : 'y'
+    }
+  } else {
+    n = e || 'xxxxxxxx-xyxx-yxxx-xxxy-xxyxxxxxxxxx'
+  }
+  return ('number' !== typeof t || t < 2 || t > 36) && (t = 16),
+    n.replace(/[xy]/g, function (e) {
+      let n = Math.random() * t | 0
+      return ('x' === e ? n : 3 & n | 8).toString(t)
+    })
 }